+++ PinionCore.Remote.Tools.Protocol.Sources.TestCommon/ISpirit1.cs
@@ -1,0 +1,10 @@
+﻿namespace PinionCore.Remote.Tools.Protocol.Sources.TestCommon
+{
+    public interface ISpirit1
+    {
+        PinionCore.Remote.Spirit<IMethodable1> Get1();
+        PinionCore.Remote.Spirit<IMethodable1> Get2(int testvalue);
+    }
+
+    
+}